Move WalletListItem and WalletImage to components
