+++ src/components/WalletImage/WalletImage.js
@@ -17,7 +17,7 @@ import PropTypes from 'prop-types'
 //   BLOCKCHAIN_LITECOIN
 // } from '@chronobank/login/network/BitcoinProvider'
 // import { BLOCKCHAIN_NEM } from '@chronobank/core/dao/NemDAO'
-import { indicators } from '../../../../../common/constants/globals'
+import { indicators } from '../../common/constants/globals'
 import {
   coin_ethereum_small,
   coin_bitcoin_cash_small,
@@ -31,7 +31,7 @@ import {
   coin_bitcoin_big,
   coin_litecoin_big,
   coin_nem_big,
-} from '../../../../../images'
+} from '../../images'
 import styles from './WalletImageStyles'
