+++ src/components/WalletImage/index.js
