+++ src/components/WalletListItem/WalletListItemContainer.js
