+++ src/components/WalletListItem/WalletListItemStyles.js
@@ -4,7 +4,7 @@
  */
 
 import { StyleSheet } from 'react-native'
-import colors from '../../../../common/colors'
+import colors from '../../common/colors'
 
 export default StyleSheet.create({
   container: {
